Move exception models into separate file
